+++ provision/env.sh
@@ -1,11 +1,12 @@
+#!/usr/bin/env bash
 # Environment-specific settings used by the provisioning process
 
 ## EXAMPLE FILE ONLY ##
 
 # Required
-DB_PASS='password' # do not use this password
+DB_PASS='password'  # do not use this password
 
 # Optional
 PUBLIC_KEY=''
-TIMEZONE='Australia/Sydney'
+TIME_ZONE='Australia/Sydney'
 DEBUG=1