+++ provision/project.sh
@@ -1,6 +1,14 @@
+#!/usr/bin/env bash
 # Put any project-specific provisioning steps here
 
-# Available variables:
-#PROJECT_NAME="$1"
-#BUILD_MODE="$2"
-#DEBUG="$3"
+# The following provisioning settings are available:
+#  - $PROJECT_NAME: The name of the project
+#  - $BUILD_MODE: The mode the provisioning is running in. One of "project" or "app"
+#  - $DEBUG: 1 if provisioning a development environment, 0 if provisioning a production environment.
+#  - $APP_DIR: The base app directory (/opt/app/)
+#  - $SRC_DIR: The directory containing the project source code (/opt/app/src/)
+#  - $PROVISION_DIR: The directory containing the project provisioning scripts (/opt/app/src/provision/)
+# See http://vagrant-django.readthedocs.io/ for more info on these settings.
+
+# Uncomment the following line to source these variables.
+#source /tmp/vagrant_provision_settings.sh