+++ CHANGELOG.rst
@@ -3,8 +3,12 @@ Change Log
 ==========
 
 0.4
-=====
+===
 
+* Moved custom scripts from ``provision/scripts/bin`` to``provision/user/bin``.
+* Moved shared folder to ``/opt/app/src``.
+* Moved other important directories under ``/opt/app``. This is now the home of everything related to the project.
+* Switched to using the "webmaster" user, created during the provisioning process, as the SSH user. The custom public key installed during provisioning is now installed for this user instead of "vagrant".
 * Added provisioning for supervisor.
 * Added provisioning for gunicorn as a production application server for Django, managed by supervisor.
 * Added provisioning for nginx as a reverse proxy to gunicorn, managed by supervisor.